Struct aws_sdk_apigateway::client::fluent_builders::CreateStage

pub struct CreateStage<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateStage<C>

pub async fn send(self) -> Result<CreateStageOutput, SdkError<CreateStageError>> where
C: SmithyConnector,

pub fn rest_api_id(self, input: impl Into<String>) -> Self

[Required] The string identifier of the associated RestApi.

pub fn set_rest_api_id(self, input: Option<String>) -> Self

pub fn stage_name(self, input: impl Into<String>) -> Self

[Required] The name for the Stage resource. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

pub fn set_stage_name(self, input: Option<String>) -> Self

pub fn deployment_id(self, input: impl Into<String>) -> Self

[Required] The identifier of the Deployment resource for the Stage resource.

pub fn set_deployment_id(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the Stage resource.

pub fn set_description(self, input: Option<String>) -> Self

pub fn cache_cluster_enabled(self, input: bool) -> Self

Whether cache clustering is enabled for the stage.

pub fn set_cache_cluster_enabled(self, input: Option<bool>) -> Self

pub fn cache_cluster_size(self, input: CacheClusterSize) -> Self

The stage's cache cluster size.

pub fn set_cache_cluster_size(self, input: Option<CacheClusterSize>) -> Self

pub fn variables(self, k: impl Into<String>, v: impl Into<String>) -> Self

A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

pub fn set_variables(self, input: Option<HashMap<String, String>>) -> Self

pub fn documentation_version(self, input: impl Into<String>) -> Self

The version of the associated API documentation.

pub fn set_documentation_version(self, input: Option<String>) -> Self

pub fn canary_settings(self, input: CanarySettings) -> Self

The canary deployment settings of this stage.

pub fn set_canary_settings(self, input: Option<CanarySettings>) -> Self

pub fn tracing_enabled(self, input: bool) -> Self

Specifies whether active tracing with X-ray is enabled for the Stage.

pub fn set_tracing_enabled(self, input: Option<bool>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self