Struct aws_sdk_apigateway::input::create_authorizer_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateAuthorizerInput

Implementations

impl Builder

pub fn rest_api_id(self, input: impl Into<String>) -> Self

[Required] The string identifier of the associated RestApi.

pub fn set_rest_api_id(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

[Required] The name of the authorizer.

pub fn set_name(self, input: Option<String>) -> Self

pub fn type(self, input: AuthorizerType) -> Self

[Required] The authorizer type. Valid values are TOKEN for a Lambda function using a single authorization token submitted in a custom header, REQUEST for a Lambda function using incoming request parameters, and COGNITO_USER_POOLS for using an Amazon Cognito user pool.

pub fn set_type(self, input: Option<AuthorizerType>) -> Self

pub fn provider_ar_ns(self, input: impl Into<String>) -> Self

pub fn set_provider_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn auth_type(self, input: impl Into<String>) -> Self

Optional customer-defined field, used in OpenAPI imports and exports without functional impact.

pub fn set_auth_type(self, input: Option<String>) -> Self

pub fn authorizer_uri(self, input: impl Into<String>) -> Self

Specifies the authorizer's Uniform Resource Identifier (URI). For TOKEN or REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form arn:aws:apigateway:{region}:lambda:path/{service_api}, where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations.

pub fn set_authorizer_uri(self, input: Option<String>) -> Self

pub fn authorizer_credentials(self, input: impl Into<String>) -> Self

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, specify null.

pub fn set_authorizer_credentials(self, input: Option<String>) -> Self

pub fn identity_source(self, input: impl Into<String>) -> Self

The identity source for which authorization is requested.

For a TOKEN or COGNITO_USER_POOLS authorizer, this is required and specifies the request header mapping expression for the custom header holding the authorization token submitted by the client. For example, if the token header name is Auth, the header mapping expression is method.request.header.Auth.

For the REQUEST authorizer, this is required when authorization caching is enabled. The value is a comma-separated string of one or more mapping expressions of the specified request parameters. For example, if an Auth header, a Name query string parameter are defined as identity sources, this value is method.request.header.Auth, method.request.querystring.Name. These parameters will be used to derive the authorization caching key and to perform runtime validation of the REQUEST authorizer by verifying all of the identity-related request parameters are present, not null and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function, otherwise, it returns a 401 Unauthorized response without calling the Lambda function. The valid value is a string of comma-separated mapping expressions of the specified request parameters. When the authorization caching is not enabled, this property is optional.

pub fn set_identity_source(self, input: Option<String>) -> Self

pub fn identity_validation_expression(self, input: impl Into<String>) -> Self

A validation expression for the incoming identity token. For TOKEN authorizers, this value is a regular expression. For COGNITO_USER_POOLS authorizers, API Gateway will match the aud field of the incoming token from the client against the specified regular expression. It will invoke the authorizer's Lambda function when there is a match. Otherwise, it will return a 401 Unauthorized response without calling the Lambda function. The validation expression does not apply to the REQUEST authorizer.

pub fn set_identity_validation_expression(self, input: Option<String>) -> Self

pub fn authorizer_result_ttl_in_seconds(self, input: i32) -> Self

The TTL in seconds of cached authorizer results. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway will cache authorizer responses. If this field is not set, the default value is 300. The maximum value is 3600, or 1 hour.

pub fn set_authorizer_result_ttl_in_seconds(self, input: Option<i32>) -> Self

pub fn build(self) -> Result<CreateAuthorizerInput, BuildError>

Consumes the builder and constructs a CreateAuthorizerInput