Module aws_sdk_apigateway::model

Modules

access_log_settings

See AccessLogSettings

api_key

See ApiKey

api_stage

See ApiStage

authorizer

See Authorizer

base_path_mapping

See BasePathMapping

canary_settings

See CanarySettings

client_certificate

See ClientCertificate

deployment

See Deployment

deployment_canary_settings

See DeploymentCanarySettings

documentation_part

See DocumentationPart

documentation_part_location

See DocumentationPartLocation

documentation_version

See DocumentationVersion

domain_name

See DomainName

endpoint_configuration

See EndpointConfiguration

gateway_response

See GatewayResponse

integration

See Integration

integration_response

See IntegrationResponse

method

See Method

method_response

See MethodResponse

method_setting

See MethodSetting

method_snapshot

See MethodSnapshot

model

See Model

mutual_tls_authentication

See MutualTlsAuthentication

mutual_tls_authentication_input

See MutualTlsAuthenticationInput

patch_operation

See PatchOperation

quota_settings

See QuotaSettings

request_validator

See RequestValidator

resource

See Resource

rest_api

See RestApi

sdk_configuration_property

See SdkConfigurationProperty

sdk_type

See SdkType

stage

See Stage

stage_key

See StageKey

throttle_settings

See ThrottleSettings

tls_config

See TlsConfig

usage_plan

See UsagePlan

usage_plan_key

See UsagePlanKey

vpc_link

See VpcLink

Structs

AccessLogSettings

Access log settings, including the access log format and access log destination ARN.

ApiKey

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

ApiStage

API stage name of the associated API stage in a usage plan.

Authorizer

Represents an authorization layer for methods. If enabled on a method, API Gateway will activate the authorizer when a client calls the method.

BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

CanarySettings

Configuration settings of a canary deployment.

ClientCertificate

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

Deployment

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

DeploymentCanarySettings

The input configuration for a canary deployment.

DocumentationPart

A documentation part for a targeted API entity.

DocumentationPartLocation

Specifies the target API entity to which the documentation applies.

DocumentationVersion

A snapshot of the documentation of an API.

DomainName

Represents a custom domain name as a user-friendly host name of an API (RestApi).

EndpointConfiguration

The endpoint configuration to indicate the types of endpoints an API (RestApi) or its custom domain name (DomainName) has.

GatewayResponse

A gateway response of a given response type and status code, with optional response parameters and mapping templates.

Integration

Represents an HTTP, HTTP_PROXY, AWS, AWS_PROXY, or Mock integration.

IntegrationResponse

Represents an integration response. The status code must map to an existing MethodResponse, and parameters and templates can be used to transform the back-end response.

Method

MethodResponse

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

MethodSetting

Specifies the method setting properties.

MethodSnapshot

Represents a summary of a Method resource, given a particular date and time.

Model

Represents the data structure of a method's request or response payload.

MutualTlsAuthentication

If specified, API Gateway performs two-way authentication between the client and the server. Clients must present a trusted certificate to access your custom domain name.

MutualTlsAuthenticationInput

If specified, API Gateway performs two-way authentication between the client and the server. Clients must present a trusted certificate to access your custom domain name.

PatchOperation

A single patch operation to apply to the specified resource. Please refer to http://tools.ietf.org/html/rfc6902#section-4 for an explanation of how each operation is used.

QuotaSettings

Quotas configured for a usage plan.

RequestValidator

A set of validation rules for incoming Method requests.

Resource

Represents an API resource.

RestApi

Represents a REST API.

SdkConfigurationProperty

A configuration property of an SDK type.

SdkType

A type of SDK that API Gateway can generate.

Stage

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

StageKey

A reference to a unique stage identified in the format {restApiId}/{stage}.

ThrottleSettings

The API request rate limits.

TlsConfig

UsagePlan

Represents a usage plan than can specify who can assess associated API stages with specified request limits and quotas.

UsagePlanKey

Represents a usage plan key to identify a plan customer.

VpcLink

An API Gateway VPC link for a RestApi to access resources in an Amazon Virtual Private Cloud (VPC).

Enums

ApiKeySourceType

ApiKeysFormat

AuthorizerType

The authorizer type. Valid values are TOKEN for a Lambda function using a single authorization token submitted in a custom header, REQUEST for a Lambda function using incoming request parameters, and COGNITO_USER_POOLS for using an Amazon Cognito user pool.

CacheClusterSize

Returns the size of the CacheCluster.

CacheClusterStatus

Returns the status of the CacheCluster.

ConnectionType

ContentHandlingStrategy

DocumentationPartType

DomainNameStatus

EndpointType

The endpoint type. The valid values are EDGE for edge-optimized API setup, most suitable for mobile applications; REGIONAL for regional API endpoint setup, most suitable for calling from AWS Region; and PRIVATE for private APIs.

GatewayResponseType

IntegrationType

The integration type. The valid value is HTTP for integrating an API method with an HTTP backend; AWS with any AWS service endpoints; MOCK for testing without actually invoking the backend; HTTP_PROXY for integrating with the HTTP proxy integration; AWS_PROXY for integrating with the Lambda proxy integration.

LocationStatusType

Op

PutMode

QuotaPeriodType

SecurityPolicy

UnauthorizedCacheControlHeaderStrategy

VpcLinkStatus