Struct aws_sdk_apigateway::model::method_setting::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for MethodSetting

Implementations

impl Builder

pub fn metrics_enabled(self, input: bool) -> Self

Specifies whether Amazon CloudWatch metrics are enabled for this method. The PATCH path for this setting is /{method_setting_key}/metrics/enabled, and the value is a Boolean.

pub fn set_metrics_enabled(self, input: Option<bool>) -> Self

pub fn logging_level(self, input: impl Into<String>) -> Self

Specifies the logging level for this method, which affects the log entries pushed to Amazon CloudWatch Logs. The PATCH path for this setting is /{method_setting_key}/logging/loglevel, and the available levels are OFF, ERROR, and INFO. Choose ERROR to write only error-level entries to CloudWatch Logs, or choose INFO to include all ERROR events as well as extra informational events.

pub fn set_logging_level(self, input: Option<String>) -> Self

pub fn data_trace_enabled(self, input: bool) -> Self

Specifies whether data trace logging is enabled for this method, which affects the log entries pushed to Amazon CloudWatch Logs. The PATCH path for this setting is /{method_setting_key}/logging/dataTrace, and the value is a Boolean.

pub fn set_data_trace_enabled(self, input: Option<bool>) -> Self

pub fn throttling_burst_limit(self, input: i32) -> Self

Specifies the throttling burst limit. The PATCH path for this setting is /{method_setting_key}/throttling/burstLimit, and the value is an integer.

pub fn set_throttling_burst_limit(self, input: Option<i32>) -> Self

pub fn throttling_rate_limit(self, input: f64) -> Self

Specifies the throttling rate limit. The PATCH path for this setting is /{method_setting_key}/throttling/rateLimit, and the value is a double.

pub fn set_throttling_rate_limit(self, input: Option<f64>) -> Self

pub fn caching_enabled(self, input: bool) -> Self

Specifies whether responses should be cached and returned for requests. A cache cluster must be enabled on the stage for responses to be cached. The PATCH path for this setting is /{method_setting_key}/caching/enabled, and the value is a Boolean.

pub fn set_caching_enabled(self, input: Option<bool>) -> Self

pub fn cache_ttl_in_seconds(self, input: i32) -> Self

Specifies the time to live (TTL), in seconds, for cached responses. The higher the TTL, the longer the response will be cached. The PATCH path for this setting is /{method_setting_key}/caching/ttlInSeconds, and the value is an integer.

pub fn set_cache_ttl_in_seconds(self, input: Option<i32>) -> Self

pub fn cache_data_encrypted(self, input: bool) -> Self

Specifies whether the cached responses are encrypted. The PATCH path for this setting is /{method_setting_key}/caching/dataEncrypted, and the value is a Boolean.

pub fn set_cache_data_encrypted(self, input: Option<bool>) -> Self

pub fn require_authorization_for_cache_control(self, input: bool) -> Self

Specifies whether authorization is required for a cache invalidation request. The PATCH path for this setting is /{method_setting_key}/caching/requireAuthorizationForCacheControl, and the value is a Boolean.

pub fn set_require_authorization_for_cache_control(
self,
input: Option<bool>
) -> Self

pub fn unauthorized_cache_control_header_strategy(
self,
input: UnauthorizedCacheControlHeaderStrategy
) -> Self

Specifies how to handle unauthorized requests for cache invalidation. The PATCH path for this setting is /{method_setting_key}/caching/unauthorizedCacheControlHeaderStrategy, and the available values are FAIL_WITH_403, SUCCEED_WITH_RESPONSE_HEADER, SUCCEED_WITHOUT_RESPONSE_HEADER.

pub fn set_unauthorized_cache_control_header_strategy(
self,
input: Option<UnauthorizedCacheControlHeaderStrategy>
) -> Self

pub fn build(self) -> MethodSetting

Consumes the builder and constructs a MethodSetting