Struct aws_sdk_apigateway::model::MethodSetting

#[non_exhaustive]
pub struct MethodSetting {
    pub metrics_enabled: bool,
    pub logging_level: Option<String>,
    pub data_trace_enabled: bool,
    pub throttling_burst_limit: i32,
    pub throttling_rate_limit: f64,
    pub caching_enabled: bool,
    pub cache_ttl_in_seconds: i32,
    pub cache_data_encrypted: bool,
    pub require_authorization_for_cache_control: bool,
    pub unauthorized_cache_control_header_strategy: Option<UnauthorizedCacheControlHeaderStrategy>,
}

Specifies the method setting properties.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

metrics_enabled: bool

Specifies whether Amazon CloudWatch metrics are enabled for this method. The PATCH path for this setting is /{method_setting_key}/metrics/enabled, and the value is a Boolean.

logging_level: Option<String>

Specifies the logging level for this method, which affects the log entries pushed to Amazon CloudWatch Logs. The PATCH path for this setting is /{method_setting_key}/logging/loglevel, and the available levels are OFF, ERROR, and INFO. Choose ERROR to write only error-level entries to CloudWatch Logs, or choose INFO to include all ERROR events as well as extra informational events.

data_trace_enabled: bool

Specifies whether data trace logging is enabled for this method, which affects the log entries pushed to Amazon CloudWatch Logs. The PATCH path for this setting is /{method_setting_key}/logging/dataTrace, and the value is a Boolean.

throttling_burst_limit: i32

Specifies the throttling burst limit. The PATCH path for this setting is /{method_setting_key}/throttling/burstLimit, and the value is an integer.

throttling_rate_limit: f64

Specifies the throttling rate limit. The PATCH path for this setting is /{method_setting_key}/throttling/rateLimit, and the value is a double.

caching_enabled: bool

Specifies whether responses should be cached and returned for requests. A cache cluster must be enabled on the stage for responses to be cached. The PATCH path for this setting is /{method_setting_key}/caching/enabled, and the value is a Boolean.

cache_ttl_in_seconds: i32

Specifies the time to live (TTL), in seconds, for cached responses. The higher the TTL, the longer the response will be cached. The PATCH path for this setting is /{method_setting_key}/caching/ttlInSeconds, and the value is an integer.

cache_data_encrypted: bool

Specifies whether the cached responses are encrypted. The PATCH path for this setting is /{method_setting_key}/caching/dataEncrypted, and the value is a Boolean.

require_authorization_for_cache_control: bool

Specifies whether authorization is required for a cache invalidation request. The PATCH path for this setting is /{method_setting_key}/caching/requireAuthorizationForCacheControl, and the value is a Boolean.

unauthorized_cache_control_header_strategy: Option<UnauthorizedCacheControlHeaderStrategy>

Specifies how to handle unauthorized requests for cache invalidation. The PATCH path for this setting is /{method_setting_key}/caching/unauthorizedCacheControlHeaderStrategy, and the available values are FAIL_WITH_403, SUCCEED_WITH_RESPONSE_HEADER, SUCCEED_WITHOUT_RESPONSE_HEADER.

Implementations

impl MethodSetting

pub fn builder() -> Builder

Creates a new builder-style object to manufacture MethodSetting