Module aws_sdk_apigateway::output

Modules

create_api_key_output

See CreateApiKeyOutput

create_authorizer_output

See CreateAuthorizerOutput

create_base_path_mapping_output

See CreateBasePathMappingOutput

create_deployment_output

See CreateDeploymentOutput

create_documentation_part_output

See CreateDocumentationPartOutput

create_documentation_version_output

See CreateDocumentationVersionOutput

create_domain_name_output

See CreateDomainNameOutput

create_model_output

See CreateModelOutput

create_request_validator_output

See CreateRequestValidatorOutput

create_resource_output

See CreateResourceOutput

create_rest_api_output

See CreateRestApiOutput

create_stage_output

See CreateStageOutput

create_usage_plan_key_output

See CreateUsagePlanKeyOutput

create_usage_plan_output

See CreateUsagePlanOutput

create_vpc_link_output

See CreateVpcLinkOutput

delete_api_key_output

See DeleteApiKeyOutput

delete_authorizer_output

See DeleteAuthorizerOutput

delete_base_path_mapping_output

See DeleteBasePathMappingOutput

delete_client_certificate_output

See DeleteClientCertificateOutput

delete_deployment_output

See DeleteDeploymentOutput

delete_documentation_part_output

See DeleteDocumentationPartOutput

delete_documentation_version_output

See DeleteDocumentationVersionOutput

delete_domain_name_output

See DeleteDomainNameOutput

delete_gateway_response_output

See DeleteGatewayResponseOutput

delete_integration_output

See DeleteIntegrationOutput

delete_integration_response_output

See DeleteIntegrationResponseOutput

delete_method_output

See DeleteMethodOutput

delete_method_response_output

See DeleteMethodResponseOutput

delete_model_output

See DeleteModelOutput

delete_request_validator_output

See DeleteRequestValidatorOutput

delete_resource_output

See DeleteResourceOutput

delete_rest_api_output

See DeleteRestApiOutput

delete_stage_output

See DeleteStageOutput

delete_usage_plan_key_output

See DeleteUsagePlanKeyOutput

delete_usage_plan_output

See DeleteUsagePlanOutput

delete_vpc_link_output

See DeleteVpcLinkOutput

flush_stage_authorizers_cache_output

See FlushStageAuthorizersCacheOutput

flush_stage_cache_output

See FlushStageCacheOutput

generate_client_certificate_output

See GenerateClientCertificateOutput

get_account_output

See GetAccountOutput

get_api_key_output

See GetApiKeyOutput

get_api_keys_output

See GetApiKeysOutput

get_authorizer_output

See GetAuthorizerOutput

get_authorizers_output

See GetAuthorizersOutput

get_base_path_mapping_output

See GetBasePathMappingOutput

get_base_path_mappings_output

See GetBasePathMappingsOutput

get_client_certificate_output

See GetClientCertificateOutput

get_client_certificates_output

See GetClientCertificatesOutput

get_deployment_output

See GetDeploymentOutput

get_deployments_output

See GetDeploymentsOutput

get_documentation_part_output

See GetDocumentationPartOutput

get_documentation_parts_output

See GetDocumentationPartsOutput

get_documentation_version_output

See GetDocumentationVersionOutput

get_documentation_versions_output

See GetDocumentationVersionsOutput

get_domain_name_output

See GetDomainNameOutput

get_domain_names_output

See GetDomainNamesOutput

get_export_output

See GetExportOutput

get_gateway_response_output

See GetGatewayResponseOutput

get_gateway_responses_output

See GetGatewayResponsesOutput

get_integration_output

See GetIntegrationOutput

get_integration_response_output

See GetIntegrationResponseOutput

get_method_output

See GetMethodOutput

get_method_response_output

See GetMethodResponseOutput

get_model_output

See GetModelOutput

get_model_template_output

See GetModelTemplateOutput

get_models_output

See GetModelsOutput

get_request_validator_output

See GetRequestValidatorOutput

get_request_validators_output

See GetRequestValidatorsOutput

get_resource_output

See GetResourceOutput

get_resources_output

See GetResourcesOutput

get_rest_api_output

See GetRestApiOutput

get_rest_apis_output

See GetRestApisOutput

get_sdk_output

See GetSdkOutput

get_sdk_type_output

See GetSdkTypeOutput

get_sdk_types_output

See GetSdkTypesOutput

get_stage_output

See GetStageOutput

get_stages_output

See GetStagesOutput

get_tags_output

See GetTagsOutput

get_usage_output

See GetUsageOutput

get_usage_plan_key_output

See GetUsagePlanKeyOutput

get_usage_plan_keys_output

See GetUsagePlanKeysOutput

get_usage_plan_output

See GetUsagePlanOutput

get_usage_plans_output

See GetUsagePlansOutput

get_vpc_link_output

See GetVpcLinkOutput

get_vpc_links_output

See GetVpcLinksOutput

import_api_keys_output

See ImportApiKeysOutput

import_documentation_parts_output

See ImportDocumentationPartsOutput

import_rest_api_output

See ImportRestApiOutput

put_gateway_response_output

See PutGatewayResponseOutput

put_integration_output

See PutIntegrationOutput

put_integration_response_output

See PutIntegrationResponseOutput

put_method_output

See PutMethodOutput

put_method_response_output

See PutMethodResponseOutput

put_rest_api_output

See PutRestApiOutput

tag_resource_output

See TagResourceOutput

test_invoke_authorizer_output

See TestInvokeAuthorizerOutput

test_invoke_method_output

See TestInvokeMethodOutput

untag_resource_output

See UntagResourceOutput

update_account_output

See UpdateAccountOutput

update_api_key_output

See UpdateApiKeyOutput

update_authorizer_output

See UpdateAuthorizerOutput

update_base_path_mapping_output

See UpdateBasePathMappingOutput

update_client_certificate_output

See UpdateClientCertificateOutput

update_deployment_output

See UpdateDeploymentOutput

update_documentation_part_output

See UpdateDocumentationPartOutput

update_documentation_version_output

See UpdateDocumentationVersionOutput

update_domain_name_output

See UpdateDomainNameOutput

update_gateway_response_output

See UpdateGatewayResponseOutput

update_integration_output

See UpdateIntegrationOutput

update_integration_response_output

See UpdateIntegrationResponseOutput

update_method_output

See UpdateMethodOutput

update_method_response_output

See UpdateMethodResponseOutput

update_model_output

See UpdateModelOutput

update_request_validator_output

See UpdateRequestValidatorOutput

update_resource_output

See UpdateResourceOutput

update_rest_api_output

See UpdateRestApiOutput

update_stage_output

See UpdateStageOutput

update_usage_output

See UpdateUsageOutput

update_usage_plan_output

See UpdateUsagePlanOutput

update_vpc_link_output

See UpdateVpcLinkOutput

Structs

CreateApiKeyOutput

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

CreateAuthorizerOutput

Represents an authorization layer for methods. If enabled on a method, API Gateway will activate the authorizer when a client calls the method.

CreateBasePathMappingOutput

Represents the base path that callers of the API must provide as part of the URL after the domain name.

CreateDeploymentOutput

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

CreateDocumentationPartOutput

A documentation part for a targeted API entity.

CreateDocumentationVersionOutput

A snapshot of the documentation of an API.

CreateDomainNameOutput

Represents a custom domain name as a user-friendly host name of an API (RestApi).

CreateModelOutput

Represents the data structure of a method's request or response payload.

CreateRequestValidatorOutput

A set of validation rules for incoming Method requests.

CreateResourceOutput

Represents an API resource.

CreateRestApiOutput

Represents a REST API.

CreateStageOutput

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

CreateUsagePlanKeyOutput

Represents a usage plan key to identify a plan customer.

CreateUsagePlanOutput

Represents a usage plan than can specify who can assess associated API stages with specified request limits and quotas.

CreateVpcLinkOutput

An API Gateway VPC link for a RestApi to access resources in an Amazon Virtual Private Cloud (VPC).

DeleteApiKeyOutput

DeleteAuthorizerOutput

DeleteBasePathMappingOutput

DeleteClientCertificateOutput

DeleteDeploymentOutput

DeleteDocumentationPartOutput

DeleteDocumentationVersionOutput

DeleteDomainNameOutput

DeleteGatewayResponseOutput

DeleteIntegrationOutput

DeleteIntegrationResponseOutput

DeleteMethodOutput

DeleteMethodResponseOutput

DeleteModelOutput

DeleteRequestValidatorOutput

DeleteResourceOutput

DeleteRestApiOutput

DeleteStageOutput

DeleteUsagePlanKeyOutput

DeleteUsagePlanOutput

DeleteVpcLinkOutput

FlushStageAuthorizersCacheOutput

FlushStageCacheOutput

GenerateClientCertificateOutput

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

GetAccountOutput

Represents an AWS account that is associated with API Gateway.

GetApiKeyOutput

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

GetApiKeysOutput

Represents a collection of API keys as represented by an ApiKeys resource.

GetAuthorizerOutput

Represents an authorization layer for methods. If enabled on a method, API Gateway will activate the authorizer when a client calls the method.

GetAuthorizersOutput

Represents a collection of Authorizer resources.

GetBasePathMappingOutput

Represents the base path that callers of the API must provide as part of the URL after the domain name.

GetBasePathMappingsOutput

Represents a collection of BasePathMapping resources.

GetClientCertificateOutput

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

GetClientCertificatesOutput

Represents a collection of ClientCertificate resources.

GetDeploymentOutput

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

GetDeploymentsOutput

Represents a collection resource that contains zero or more references to your existing deployments, and links that guide you on how to interact with your collection. The collection offers a paginated view of the contained deployments.

GetDocumentationPartOutput

A documentation part for a targeted API entity.

GetDocumentationPartsOutput

The collection of documentation parts of an API.

GetDocumentationVersionOutput

A snapshot of the documentation of an API.

GetDocumentationVersionsOutput

The collection of documentation snapshots of an API.

GetDomainNameOutput

Represents a custom domain name as a user-friendly host name of an API (RestApi).

GetDomainNamesOutput

Represents a collection of DomainName resources.

GetExportOutput

The binary blob response to GetExport, which contains the generated SDK.

GetGatewayResponseOutput

A gateway response of a given response type and status code, with optional response parameters and mapping templates.

GetGatewayResponsesOutput

The collection of the GatewayResponse instances of a RestApi as a responseType-to-GatewayResponse object map of key-value pairs. As such, pagination is not supported for querying this collection.

GetIntegrationOutput

Represents an HTTP, HTTP_PROXY, AWS, AWS_PROXY, or Mock integration.

GetIntegrationResponseOutput

Represents an integration response. The status code must map to an existing MethodResponse, and parameters and templates can be used to transform the back-end response.

GetMethodOutput

GetMethodResponseOutput

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

GetModelOutput

Represents the data structure of a method's request or response payload.

GetModelTemplateOutput

Represents a mapping template used to transform a payload.

GetModelsOutput

Represents a collection of Model resources.

GetRequestValidatorOutput

A set of validation rules for incoming Method requests.

GetRequestValidatorsOutput

A collection of RequestValidator resources of a given RestApi.

GetResourceOutput

Represents an API resource.

GetResourcesOutput

Represents a collection of Resource resources.

GetRestApiOutput

Represents a REST API.

GetRestApisOutput

Contains references to your APIs and links that guide you in how to interact with your collection. A collection offers a paginated view of your APIs.

GetSdkOutput

The binary blob response to GetSdk, which contains the generated SDK.

GetSdkTypeOutput

A type of SDK that API Gateway can generate.

GetSdkTypesOutput

The collection of SdkType instances.

GetStageOutput

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

GetStagesOutput

A list of Stage resources that are associated with the ApiKey resource.

GetTagsOutput

The collection of tags. Each tag element is associated with a given resource.

GetUsageOutput

Represents the usage data of a usage plan.

GetUsagePlanKeyOutput

Represents a usage plan key to identify a plan customer.

GetUsagePlanKeysOutput

Represents the collection of usage plan keys added to usage plans for the associated API keys and, possibly, other types of keys.

GetUsagePlanOutput

Represents a usage plan than can specify who can assess associated API stages with specified request limits and quotas.

GetUsagePlansOutput

Represents a collection of usage plans for an AWS account.

GetVpcLinkOutput

An API Gateway VPC link for a RestApi to access resources in an Amazon Virtual Private Cloud (VPC).

GetVpcLinksOutput

The collection of VPC links under the caller's account in a region.

ImportApiKeysOutput

The identifier of an ApiKey used in a UsagePlan.

ImportDocumentationPartsOutput

A collection of the imported DocumentationPart identifiers.

ImportRestApiOutput

Represents a REST API.

PutGatewayResponseOutput

A gateway response of a given response type and status code, with optional response parameters and mapping templates.

PutIntegrationOutput

Represents an HTTP, HTTP_PROXY, AWS, AWS_PROXY, or Mock integration.

PutIntegrationResponseOutput

Represents an integration response. The status code must map to an existing MethodResponse, and parameters and templates can be used to transform the back-end response.

PutMethodOutput

PutMethodResponseOutput

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

PutRestApiOutput

Represents a REST API.

TagResourceOutput

TestInvokeAuthorizerOutput

Represents the response of the test invoke request for a custom Authorizer

TestInvokeMethodOutput

Represents the response of the test invoke request in the HTTP method.

UntagResourceOutput

UpdateAccountOutput

Represents an AWS account that is associated with API Gateway.

UpdateApiKeyOutput

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

UpdateAuthorizerOutput

Represents an authorization layer for methods. If enabled on a method, API Gateway will activate the authorizer when a client calls the method.

UpdateBasePathMappingOutput

Represents the base path that callers of the API must provide as part of the URL after the domain name.

UpdateClientCertificateOutput

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

UpdateDeploymentOutput

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

UpdateDocumentationPartOutput

A documentation part for a targeted API entity.

UpdateDocumentationVersionOutput

A snapshot of the documentation of an API.

UpdateDomainNameOutput

Represents a custom domain name as a user-friendly host name of an API (RestApi).

UpdateGatewayResponseOutput

A gateway response of a given response type and status code, with optional response parameters and mapping templates.

UpdateIntegrationOutput

Represents an HTTP, HTTP_PROXY, AWS, AWS_PROXY, or Mock integration.

UpdateIntegrationResponseOutput

Represents an integration response. The status code must map to an existing MethodResponse, and parameters and templates can be used to transform the back-end response.

UpdateMethodOutput

UpdateMethodResponseOutput

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

UpdateModelOutput

Represents the data structure of a method's request or response payload.

UpdateRequestValidatorOutput

A set of validation rules for incoming Method requests.

UpdateResourceOutput

Represents an API resource.

UpdateRestApiOutput

Represents a REST API.

UpdateStageOutput

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

UpdateUsageOutput

Represents the usage data of a usage plan.

UpdateUsagePlanOutput

Represents a usage plan than can specify who can assess associated API stages with specified request limits and quotas.

UpdateVpcLinkOutput

An API Gateway VPC link for a RestApi to access resources in an Amazon Virtual Private Cloud (VPC).