Struct aws_sdk_apigateway::output::get_integration_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetIntegrationOutput

Implementations

impl Builder

pub fn type(self, input: IntegrationType) -> Self

Specifies an API method integration type. The valid value is one of the following:

AWS: for integrating the API method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration.

AWS_PROXY: for integrating the API method request with the Lambda function-invoking action with the client request passed through as-is. This integration is also referred to as the Lambda proxy integration.

HTTP: for integrating the API method request with an HTTP endpoint, including a private HTTP endpoint within a VPC. This integration is also referred to as the HTTP custom integration.

HTTP_PROXY: for integrating the API method request with an HTTP endpoint, including a private HTTP endpoint within a VPC, with the client request passed through as-is. This is also referred to as the HTTP proxy integration.

MOCK: for integrating the API method request with API Gateway as a "loop-back" endpoint without invoking any backend.

For the HTTP and HTTP proxy integrations, each integration can specify a protocol (http/https), port and path. Standard 80 and 443 ports are supported as well as custom ports above 1024. An HTTP or HTTP proxy integration with a connectionType of VPC_LINK is referred to as a private integration and uses a VpcLink to connect API Gateway to a network load balancer of a VPC.

pub fn set_type(self, input: Option<IntegrationType>) -> Self

pub fn http_method(self, input: impl Into<String>) -> Self

Specifies the integration's HTTP method type.

pub fn set_http_method(self, input: Option<String>) -> Self

pub fn uri(self, input: impl Into<String>) -> Self

Specifies Uniform Resource Identifier (URI) of the integration endpoint.

For HTTP or HTTP_PROXY integrations, the URI must be a fully formed, encoded HTTP(S) URL according to the RFC-3986 specification, for either standard integration, where connectionType is not VPC_LINK, or private integration, where connectionType is VPC_LINK. For a private HTTP integration, the URI is not used for routing.

For AWS or AWS_PROXY integrations, the URI is of the form arn:aws:apigateway:{region}:{subdomain.service|service}:path|action/{service_api}. Here, {Region} is the API Gateway region (e.g., us-east-1); {service} is the name of the integrated AWS service (e.g., s3); and {subdomain} is a designated subdomain supported by certain AWS service for fast host-name lookup. action can be used for an AWS service action-based API, using an Action={name}&{p1}={v1}&p2={v2}... query string. The ensuing {service_api} refers to a supported action {name} plus any required input parameters. Alternatively, path can be used for an AWS service path-based API. The ensuing service_api refers to the path to an AWS service resource, including the region of the integrated AWS service, if applicable. For example, for integration with the S3 API of GetObject, the uri can be either arn:aws:apigateway:us-west-2:s3:action/GetObject&Bucket={bucket}&Key={key} or arn:aws:apigateway:us-west-2:s3:path/{bucket}/{key}

pub fn set_uri(self, input: Option<String>) -> Self

pub fn connection_type(self, input: ConnectionType) -> Self

The type of the network connection to the integration endpoint. The valid value is INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and a network load balancer in a VPC. The default value is INTERNET.

pub fn set_connection_type(self, input: Option<ConnectionType>) -> Self

pub fn connection_id(self, input: impl Into<String>) -> Self

The (id) of the VpcLink used for the integration when connectionType=VPC_LINK and undefined, otherwise.

pub fn set_connection_id(self, input: Option<String>) -> Self

pub fn credentials(self, input: impl Into<String>) -> Self

Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam::\*:user/\*. To use resource-based permissions on supported AWS services, specify null.

pub fn set_credentials(self, input: Option<String>) -> Self

pub fn request_parameters(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_request_parameters(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn request_templates(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_request_templates(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn passthrough_behavior(self, input: impl Into<String>) -> Self

Specifies how the method request body of an unmapped content type will be passed through the integration request to the back end without transformation. A content type is unmapped if no mapping template is defined in the integration or the content type does not match any of the mapped content types, as specified in requestTemplates. The valid value is one of the following:

WHEN_NO_MATCH: passes the method request body through the integration request to the back end without transformation when the method request content type does not match any content type associated with the mapping templates defined in the integration request.

WHEN_NO_TEMPLATES: passes the method request body through the integration request to the back end without transformation when no mapping template is defined in the integration request. If a template is defined when this option is selected, the method request of an unmapped content-type will be rejected with an HTTP 415 Unsupported Media Type response.

NEVER: rejects the method request with an HTTP 415 Unsupported Media Type response when either the method request content type does not match any content type associated with the mapping templates defined in the integration request or no mapping template is defined in the integration request.

pub fn set_passthrough_behavior(self, input: Option<String>) -> Self

pub fn content_handling(self, input: ContentHandlingStrategy) -> Self

Specifies how to handle request payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a request payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a request payload from a binary blob to a Base64-encoded string.

If this property is not defined, the request payload will be passed through from the method request to integration request without modification, provided that the passthroughBehavior is configured to support payload pass-through.

pub fn set_content_handling(
self,
input: Option<ContentHandlingStrategy>
) -> Self

pub fn timeout_in_millis(self, input: i32) -> Self

Custom timeout between 50 and 29,000 milliseconds. The default value is 29,000 milliseconds or 29 seconds.

pub fn set_timeout_in_millis(self, input: Option<i32>) -> Self

pub fn cache_namespace(self, input: impl Into<String>) -> Self

Specifies a group of related cached parameters. By default, API Gateway uses the resource ID as the cacheNamespace. You can specify the same cacheNamespace across resources to return the same cached data for requests to different resources.

pub fn set_cache_namespace(self, input: Option<String>) -> Self

pub fn cache_key_parameters(self, input: impl Into<String>) -> Self

pub fn set_cache_key_parameters(self, input: Option<Vec<String>>) -> Self

pub fn integration_responses(
self,
k: impl Into<String>,
v: impl Into<IntegrationResponse>
) -> Self

pub fn set_integration_responses(
self,
input: Option<HashMap<String, IntegrationResponse>>
) -> Self

pub fn tls_config(self, input: TlsConfig) -> Self

Specifies the TLS configuration for an integration.

pub fn set_tls_config(self, input: Option<TlsConfig>) -> Self

pub fn build(self) -> GetIntegrationOutput

Consumes the builder and constructs a GetIntegrationOutput