Struct aws_sdk_apigateway::output::get_stage_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for GetStageOutput

Implementations

impl Builder

pub fn deployment_id(self, input: impl Into<String>) -> Self

The identifier of the Deployment that the stage points to.

pub fn set_deployment_id(self, input: Option<String>) -> Self

pub fn client_certificate_id(self, input: impl Into<String>) -> Self

The identifier of a client certificate for an API stage.

pub fn set_client_certificate_id(self, input: Option<String>) -> Self

pub fn stage_name(self, input: impl Into<String>) -> Self

The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to API Gateway. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

pub fn set_stage_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The stage's description.

pub fn set_description(self, input: Option<String>) -> Self

pub fn cache_cluster_enabled(self, input: bool) -> Self

Specifies whether a cache cluster is enabled for the stage.

pub fn set_cache_cluster_enabled(self, input: Option<bool>) -> Self

pub fn cache_cluster_size(self, input: CacheClusterSize) -> Self

The size of the cache cluster for the stage, if enabled.

pub fn set_cache_cluster_size(self, input: Option<CacheClusterSize>) -> Self

pub fn cache_cluster_status(self, input: CacheClusterStatus) -> Self

The status of the cache cluster for the stage, if enabled.

pub fn set_cache_cluster_status(self, input: Option<CacheClusterStatus>) -> Self

pub fn method_settings(
self,
k: impl Into<String>,
v: impl Into<MethodSetting>
) -> Self

pub fn set_method_settings(
self,
input: Option<HashMap<String, MethodSetting>>
) -> Self

pub fn variables(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_variables(self, input: Option<HashMap<String, String>>) -> Self

pub fn documentation_version(self, input: impl Into<String>) -> Self

The version of the associated API documentation.

pub fn set_documentation_version(self, input: Option<String>) -> Self

pub fn access_log_settings(self, input: AccessLogSettings) -> Self

Settings for logging access in this stage.

pub fn set_access_log_settings(self, input: Option<AccessLogSettings>) -> Self

pub fn canary_settings(self, input: CanarySettings) -> Self

Settings for the canary deployment in this stage.

pub fn set_canary_settings(self, input: Option<CanarySettings>) -> Self

pub fn tracing_enabled(self, input: bool) -> Self

Specifies whether active tracing with X-ray is enabled for the Stage.

pub fn set_tracing_enabled(self, input: Option<bool>) -> Self

pub fn web_acl_arn(self, input: impl Into<String>) -> Self

The ARN of the WebAcl associated with the Stage.

pub fn set_web_acl_arn(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn created_date(self, input: Instant) -> Self

The timestamp when the stage was created.

pub fn set_created_date(self, input: Option<Instant>) -> Self

pub fn last_updated_date(self, input: Instant) -> Self

The timestamp when the stage last updated.

pub fn set_last_updated_date(self, input: Option<Instant>) -> Self

pub fn build(self) -> GetStageOutput

Consumes the builder and constructs a GetStageOutput