Module aws_sdk_apigateway::error

Modules

bad_request_error
See BadRequestError

conflict_error
See ConflictError

limit_exceeded_error
See LimitExceededError

not_found_error
See NotFoundError

service_unavailable_error
See ServiceUnavailableError

too_many_requests_error
See TooManyRequestsError

unauthorized_error
See UnauthorizedError

Structs

BadRequestError
The submitted request is not valid, for example, the input is incomplete or incorrect. See the accompanying error message for details.

ConflictError
The request configuration has conflicts. For details, see the accompanying error message.

CreateApiKeyError

CreateAuthorizerError

CreateBasePathMappingError

CreateDeploymentError

CreateDocumentationPartError

CreateDocumentationVersionError

CreateDomainNameError

CreateModelError

CreateRequestValidatorError

CreateResourceError

CreateRestApiError

CreateStageError

CreateUsagePlanError

CreateUsagePlanKeyError

CreateVpcLinkError

DeleteApiKeyError

DeleteAuthorizerError

DeleteBasePathMappingError

DeleteClientCertificateError

DeleteDeploymentError

DeleteDocumentationPartError

DeleteDocumentationVersionError

DeleteDomainNameError

DeleteGatewayResponseError

DeleteIntegrationError

DeleteIntegrationResponseError

DeleteMethodError

DeleteMethodResponseError

DeleteModelError

DeleteRequestValidatorError

DeleteResourceError

DeleteRestApiError

DeleteStageError

DeleteUsagePlanError

DeleteUsagePlanKeyError

DeleteVpcLinkError

FlushStageAuthorizersCacheError

FlushStageCacheError

GenerateClientCertificateError

GetAccountError

GetApiKeyError

GetApiKeysError

GetAuthorizerError

GetAuthorizersError

GetBasePathMappingError

GetBasePathMappingsError

GetClientCertificateError

GetClientCertificatesError

GetDeploymentError

GetDeploymentsError

GetDocumentationPartError

GetDocumentationPartsError

GetDocumentationVersionError

GetDocumentationVersionsError

GetDomainNameError

GetDomainNamesError

GetExportError

GetGatewayResponseError

GetGatewayResponsesError

GetIntegrationError

GetIntegrationResponseError

GetMethodError

GetMethodResponseError

GetModelError

GetModelTemplateError

GetModelsError

GetRequestValidatorError

GetRequestValidatorsError

GetResourceError

GetResourcesError

GetRestApiError

GetRestApisError

GetSdkError

GetSdkTypeError

GetSdkTypesError

GetStageError

GetStagesError

GetTagsError

GetUsageError

GetUsagePlanError

GetUsagePlanKeyError

GetUsagePlanKeysError

GetUsagePlansError

GetVpcLinkError

GetVpcLinksError

ImportApiKeysError

ImportDocumentationPartsError

ImportRestApiError

LimitExceededError
The request exceeded the rate limit. Retry after the specified time period.

NotFoundError
The requested resource is not found. Make sure that the request URI is correct.

PutGatewayResponseError

PutIntegrationError

PutIntegrationResponseError

PutMethodError

PutMethodResponseError

PutRestApiError

ServiceUnavailableError
The requested service is not available. For details see the accompanying error message. Retry after the specified time period.

TagResourceError

TestInvokeAuthorizerError

TestInvokeMethodError

TooManyRequestsError
The request has reached its throttling limit. Retry after the specified time period.

UnauthorizedError
The request is denied because the caller has insufficient permissions.

UntagResourceError

UpdateAccountError

UpdateApiKeyError

UpdateAuthorizerError

UpdateBasePathMappingError

UpdateClientCertificateError

UpdateDeploymentError

UpdateDocumentationPartError

UpdateDocumentationVersionError

UpdateDomainNameError

UpdateGatewayResponseError

UpdateIntegrationError

UpdateIntegrationResponseError

UpdateMethodError

UpdateMethodResponseError

UpdateModelError

UpdateRequestValidatorError

UpdateResourceError

UpdateRestApiError

UpdateStageError

UpdateUsageError

UpdateUsagePlanError

UpdateVpcLinkError

Enums

CreateApiKeyErrorKind

CreateAuthorizerErrorKind

CreateBasePathMappingErrorKind

CreateDeploymentErrorKind

CreateDocumentationPartErrorKind

CreateDocumentationVersionErrorKind

CreateDomainNameErrorKind

CreateModelErrorKind

CreateRequestValidatorErrorKind

CreateResourceErrorKind

CreateRestApiErrorKind

CreateStageErrorKind

CreateUsagePlanErrorKind

CreateUsagePlanKeyErrorKind

CreateVpcLinkErrorKind

DeleteApiKeyErrorKind

DeleteAuthorizerErrorKind

DeleteBasePathMappingErrorKind

DeleteClientCertificateErrorKind

DeleteDeploymentErrorKind

DeleteDocumentationPartErrorKind

DeleteDocumentationVersionErrorKind

DeleteDomainNameErrorKind

DeleteGatewayResponseErrorKind

DeleteIntegrationErrorKind

DeleteIntegrationResponseErrorKind

DeleteMethodErrorKind

DeleteMethodResponseErrorKind

DeleteModelErrorKind

DeleteRequestValidatorErrorKind

DeleteResourceErrorKind

DeleteRestApiErrorKind

DeleteStageErrorKind

DeleteUsagePlanErrorKind

DeleteUsagePlanKeyErrorKind

DeleteVpcLinkErrorKind

FlushStageAuthorizersCacheErrorKind

FlushStageCacheErrorKind

GenerateClientCertificateErrorKind

GetAccountErrorKind

GetApiKeyErrorKind

GetApiKeysErrorKind

GetAuthorizerErrorKind

GetAuthorizersErrorKind

GetBasePathMappingErrorKind

GetBasePathMappingsErrorKind

GetClientCertificateErrorKind

GetClientCertificatesErrorKind

GetDeploymentErrorKind

GetDeploymentsErrorKind

GetDocumentationPartErrorKind

GetDocumentationPartsErrorKind

GetDocumentationVersionErrorKind

GetDocumentationVersionsErrorKind

GetDomainNameErrorKind

GetDomainNamesErrorKind

GetExportErrorKind

GetGatewayResponseErrorKind

GetGatewayResponsesErrorKind

GetIntegrationErrorKind

GetIntegrationResponseErrorKind

GetMethodErrorKind

GetMethodResponseErrorKind

GetModelErrorKind

GetModelTemplateErrorKind

GetModelsErrorKind

GetRequestValidatorErrorKind

GetRequestValidatorsErrorKind

GetResourceErrorKind

GetResourcesErrorKind

GetRestApiErrorKind

GetRestApisErrorKind

GetSdkErrorKind

GetSdkTypeErrorKind

GetSdkTypesErrorKind

GetStageErrorKind

GetStagesErrorKind

GetTagsErrorKind

GetUsageErrorKind

GetUsagePlanErrorKind

GetUsagePlanKeyErrorKind

GetUsagePlanKeysErrorKind

GetUsagePlansErrorKind

GetVpcLinkErrorKind

GetVpcLinksErrorKind

ImportApiKeysErrorKind

ImportDocumentationPartsErrorKind

ImportRestApiErrorKind

PutGatewayResponseErrorKind

PutIntegrationErrorKind

PutIntegrationResponseErrorKind

PutMethodErrorKind

PutMethodResponseErrorKind

PutRestApiErrorKind

TagResourceErrorKind

TestInvokeAuthorizerErrorKind

TestInvokeMethodErrorKind

UntagResourceErrorKind

UpdateAccountErrorKind

UpdateApiKeyErrorKind

UpdateAuthorizerErrorKind

UpdateBasePathMappingErrorKind

UpdateClientCertificateErrorKind

UpdateDeploymentErrorKind

UpdateDocumentationPartErrorKind

UpdateDocumentationVersionErrorKind

UpdateDomainNameErrorKind

UpdateGatewayResponseErrorKind

UpdateIntegrationErrorKind

UpdateIntegrationResponseErrorKind

UpdateMethodErrorKind

UpdateMethodResponseErrorKind

UpdateModelErrorKind

UpdateRequestValidatorErrorKind

UpdateResourceErrorKind

UpdateRestApiErrorKind

UpdateStageErrorKind

UpdateUsageErrorKind

UpdateUsagePlanErrorKind

UpdateVpcLinkErrorKind