Module aws_sdk_apigateway::input

Modules

create_api_key_input
See CreateApiKeyInput

create_authorizer_input
See CreateAuthorizerInput

create_base_path_mapping_input
See CreateBasePathMappingInput

create_deployment_input
See CreateDeploymentInput

create_documentation_part_input
See CreateDocumentationPartInput

create_documentation_version_input
See CreateDocumentationVersionInput

create_domain_name_input
See CreateDomainNameInput

create_model_input
See CreateModelInput

create_request_validator_input
See CreateRequestValidatorInput

create_resource_input
See CreateResourceInput

create_rest_api_input
See CreateRestApiInput

create_stage_input
See CreateStageInput

create_usage_plan_input
See CreateUsagePlanInput

create_usage_plan_key_input
See CreateUsagePlanKeyInput

create_vpc_link_input
See CreateVpcLinkInput

delete_api_key_input
See DeleteApiKeyInput

delete_authorizer_input
See DeleteAuthorizerInput

delete_base_path_mapping_input
See DeleteBasePathMappingInput

delete_client_certificate_input
See DeleteClientCertificateInput

delete_deployment_input
See DeleteDeploymentInput

delete_documentation_part_input
See DeleteDocumentationPartInput

delete_documentation_version_input
See DeleteDocumentationVersionInput

delete_domain_name_input
See DeleteDomainNameInput

delete_gateway_response_input
See DeleteGatewayResponseInput

delete_integration_input
See DeleteIntegrationInput

delete_integration_response_input
See DeleteIntegrationResponseInput

delete_method_input
See DeleteMethodInput

delete_method_response_input
See DeleteMethodResponseInput

delete_model_input
See DeleteModelInput

delete_request_validator_input
See DeleteRequestValidatorInput

delete_resource_input
See DeleteResourceInput

delete_rest_api_input
See DeleteRestApiInput

delete_stage_input
See DeleteStageInput

delete_usage_plan_input
See DeleteUsagePlanInput

delete_usage_plan_key_input
See DeleteUsagePlanKeyInput

delete_vpc_link_input
See DeleteVpcLinkInput

flush_stage_authorizers_cache_input
See FlushStageAuthorizersCacheInput

flush_stage_cache_input
See FlushStageCacheInput

generate_client_certificate_input
See GenerateClientCertificateInput

get_account_input
See GetAccountInput

get_api_key_input
See GetApiKeyInput

get_api_keys_input
See GetApiKeysInput

get_authorizer_input
See GetAuthorizerInput

get_authorizers_input
See GetAuthorizersInput

get_base_path_mapping_input
See GetBasePathMappingInput

get_base_path_mappings_input
See GetBasePathMappingsInput

get_client_certificate_input
See GetClientCertificateInput

get_client_certificates_input
See GetClientCertificatesInput

get_deployment_input
See GetDeploymentInput

get_deployments_input
See GetDeploymentsInput

get_documentation_part_input
See GetDocumentationPartInput

get_documentation_parts_input
See GetDocumentationPartsInput

get_documentation_version_input
See GetDocumentationVersionInput

get_documentation_versions_input
See GetDocumentationVersionsInput

get_domain_name_input
See GetDomainNameInput

get_domain_names_input
See GetDomainNamesInput

get_export_input
See GetExportInput

get_gateway_response_input
See GetGatewayResponseInput

get_gateway_responses_input
See GetGatewayResponsesInput

get_integration_input
See GetIntegrationInput

get_integration_response_input
See GetIntegrationResponseInput

get_method_input
See GetMethodInput

get_method_response_input
See GetMethodResponseInput

get_model_input
See GetModelInput

get_model_template_input
See GetModelTemplateInput

get_models_input
See GetModelsInput

get_request_validator_input
See GetRequestValidatorInput

get_request_validators_input
See GetRequestValidatorsInput

get_resource_input
See GetResourceInput

get_resources_input
See GetResourcesInput

get_rest_api_input
See GetRestApiInput

get_rest_apis_input
See GetRestApisInput

get_sdk_input
See GetSdkInput

get_sdk_type_input
See GetSdkTypeInput

get_sdk_types_input
See GetSdkTypesInput

get_stage_input
See GetStageInput

get_stages_input
See GetStagesInput

get_tags_input
See GetTagsInput

get_usage_input
See GetUsageInput

get_usage_plan_input
See GetUsagePlanInput

get_usage_plan_key_input
See GetUsagePlanKeyInput

get_usage_plan_keys_input
See GetUsagePlanKeysInput

get_usage_plans_input
See GetUsagePlansInput

get_vpc_link_input
See GetVpcLinkInput

get_vpc_links_input
See GetVpcLinksInput

import_api_keys_input
See ImportApiKeysInput

import_documentation_parts_input
See ImportDocumentationPartsInput

import_rest_api_input
See ImportRestApiInput

put_gateway_response_input
See PutGatewayResponseInput

put_integration_input
See PutIntegrationInput

put_integration_response_input
See PutIntegrationResponseInput

put_method_input
See PutMethodInput

put_method_response_input
See PutMethodResponseInput

put_rest_api_input
See PutRestApiInput

tag_resource_input
See TagResourceInput

test_invoke_authorizer_input
See TestInvokeAuthorizerInput

test_invoke_method_input
See TestInvokeMethodInput

untag_resource_input
See UntagResourceInput

update_account_input
See UpdateAccountInput

update_api_key_input
See UpdateApiKeyInput

update_authorizer_input
See UpdateAuthorizerInput

update_base_path_mapping_input
See UpdateBasePathMappingInput

update_client_certificate_input
See UpdateClientCertificateInput

update_deployment_input
See UpdateDeploymentInput

update_documentation_part_input
See UpdateDocumentationPartInput

update_documentation_version_input
See UpdateDocumentationVersionInput

update_domain_name_input
See UpdateDomainNameInput

update_gateway_response_input
See UpdateGatewayResponseInput

update_integration_input
See UpdateIntegrationInput

update_integration_response_input
See UpdateIntegrationResponseInput

update_method_input
See UpdateMethodInput

update_method_response_input
See UpdateMethodResponseInput

update_model_input
See UpdateModelInput

update_request_validator_input
See UpdateRequestValidatorInput

update_resource_input
See UpdateResourceInput

update_rest_api_input
See UpdateRestApiInput

update_stage_input
See UpdateStageInput

update_usage_input
See UpdateUsageInput

update_usage_plan_input
See UpdateUsagePlanInput

update_vpc_link_input
See UpdateVpcLinkInput

Structs

CreateApiKeyInput
Request to create an ApiKey resource.

CreateAuthorizerInput
Request to add a new Authorizer to an existing RestApi resource.

CreateBasePathMappingInput
Requests API Gateway to create a new BasePathMapping resource.

CreateDeploymentInput
Requests API Gateway to create a Deployment resource.

CreateDocumentationPartInput
Creates a new documentation part of a given API.

CreateDocumentationVersionInput
Creates a new documentation version of a given API.

CreateDomainNameInput
A request to create a new domain name.

CreateModelInput
Request to add a new Model to an existing RestApi resource.

CreateRequestValidatorInput
Creates a RequestValidator of a given RestApi.

CreateResourceInput
Requests API Gateway to create a Resource resource.

CreateRestApiInput
The POST Request to add a new RestApi resource to your collection.

CreateStageInput
Requests API Gateway to create a Stage resource.

CreateUsagePlanInput
The POST request to create a usage plan with the name, description, throttle limits and quota limits, as well as the associated API stages, specified in the payload.

CreateUsagePlanKeyInput
The POST request to create a usage plan key for adding an existing API key to a usage plan.

CreateVpcLinkInput
Creates a VPC link, under the caller's account in a selected region, in an asynchronous operation that typically takes 2-4 minutes to complete and become operational. The caller must have permissions to create and update VPC Endpoint services.

DeleteApiKeyInput
A request to delete the ApiKey resource.

DeleteAuthorizerInput
Request to delete an existing Authorizer resource.

DeleteBasePathMappingInput
A request to delete the BasePathMapping resource.

DeleteClientCertificateInput
A request to delete the ClientCertificate resource.

DeleteDeploymentInput
Requests API Gateway to delete a Deployment resource.

DeleteDocumentationPartInput
Deletes an existing documentation part of an API.

DeleteDocumentationVersionInput
Deletes an existing documentation version of an API.

DeleteDomainNameInput
A request to delete the DomainName resource.

DeleteGatewayResponseInput
Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

DeleteIntegrationInput
Represents a delete integration request.

DeleteIntegrationResponseInput
Represents a delete integration response request.

DeleteMethodInput
Request to delete an existing Method resource.

DeleteMethodResponseInput
A request to delete an existing MethodResponse resource.

DeleteModelInput
Request to delete an existing model in an existing RestApi resource.

DeleteRequestValidatorInput
Deletes a specified RequestValidator of a given RestApi.

DeleteResourceInput
Request to delete a Resource.

DeleteRestApiInput
Request to delete the specified API from your collection.

DeleteStageInput
Requests API Gateway to delete a Stage resource.

DeleteUsagePlanInput
The DELETE request to delete a usage plan of a given plan Id.

DeleteUsagePlanKeyInput
The DELETE request to delete a usage plan key and remove the underlying API key from the associated usage plan.

DeleteVpcLinkInput
Deletes an existing VpcLink of a specified identifier.

FlushStageAuthorizersCacheInput
Request to flush authorizer cache entries on a specified stage.

FlushStageCacheInput
Requests API Gateway to flush a stage's cache.

GenerateClientCertificateInput
A request to generate a ClientCertificate resource.

GetAccountInput
Requests API Gateway to get information about the current Account resource.

GetApiKeyInput
A request to get information about the current ApiKey resource.

GetApiKeysInput
A request to get information about the current ApiKeys resource.

GetAuthorizerInput
Request to describe an existing Authorizer resource.

GetAuthorizersInput
Request to describe an existing Authorizers resource.

GetBasePathMappingInput
Request to describe a BasePathMapping resource.

GetBasePathMappingsInput
A request to get information about a collection of BasePathMapping resources.

GetClientCertificateInput
A request to get information about the current ClientCertificate resource.

GetClientCertificatesInput
A request to get information about a collection of ClientCertificate resources.

GetDeploymentInput
Requests API Gateway to get information about a Deployment resource.

GetDeploymentsInput
Requests API Gateway to get information about a Deployments collection.

GetDocumentationPartInput
Gets a specified documentation part of a given API.

GetDocumentationPartsInput
Gets the documentation parts of an API. The result may be filtered by the type, name, or path of API entities (targets).

GetDocumentationVersionInput
Gets a documentation snapshot of an API.

GetDocumentationVersionsInput
Gets the documentation versions of an API.

GetDomainNameInput
Request to get the name of a DomainName resource.

GetDomainNamesInput
Request to describe a collection of DomainName resources.

GetExportInput
Request a new export of a RestApi for a particular Stage.

GetGatewayResponseInput
Gets a GatewayResponse of a specified response type on the given RestApi.

GetGatewayResponsesInput
Gets the GatewayResponses collection on the given RestApi. If an API developer has not added any definitions for gateway responses, the result will be the API Gateway-generated default GatewayResponses collection for the supported response types.

GetIntegrationInput
Represents a request to get the integration configuration.

GetIntegrationResponseInput
Represents a get integration response request.

GetMethodInput
Request to describe an existing Method resource.

GetMethodResponseInput
Request to describe a MethodResponse resource.

GetModelInput
Request to list information about a model in an existing RestApi resource.

GetModelTemplateInput
Request to generate a sample mapping template used to transform the payload.

GetModelsInput
Request to list existing Models defined for a RestApi resource.

GetRequestValidatorInput
Gets a RequestValidator of a given RestApi.

GetRequestValidatorsInput
Gets the RequestValidators collection of a given RestApi.

GetResourceInput
Request to list information about a resource.

GetResourcesInput
Request to list information about a collection of resources.

GetRestApiInput
The GET request to list an existing RestApi defined for your collection.

GetRestApisInput
The GET request to list existing RestApis defined for your collection.

GetSdkInput
Request a new generated client SDK for a RestApi and Stage.

GetSdkTypeInput
Get an SdkType instance.

GetSdkTypesInput
Get the SdkTypes collection.

GetStageInput
Requests API Gateway to get information about a Stage resource.

GetStagesInput
Requests API Gateway to get information about one or more Stage resources.

GetTagsInput
Gets the Tags collection for a given resource.

GetUsageInput
The GET request to get the usage data of a usage plan in a specified time interval.

GetUsagePlanInput
The GET request to get a usage plan of a given plan identifier.

GetUsagePlanKeyInput
The GET request to get a usage plan key of a given key identifier.

GetUsagePlanKeysInput
The GET request to get all the usage plan keys representing the API keys added to a specified usage plan.

GetUsagePlansInput
The GET request to get all the usage plans of the caller's account.

GetVpcLinkInput
Gets a specified VPC link under the caller's account in a region.

GetVpcLinksInput
Gets the VpcLinks collection under the caller's account in a selected region.

ImportApiKeysInput
The POST request to import API keys from an external source, such as a CSV-formatted file.

ImportDocumentationPartsInput
Import documentation parts from an external (e.g., OpenAPI) definition file.

ImportRestApiInput
A POST request to import an API to API Gateway using an input of an API definition file.

PutGatewayResponseInput
Creates a customization of a GatewayResponse of a specified response type and status code on the given RestApi.

PutIntegrationInput
Sets up a method's integration.

PutIntegrationResponseInput
Represents a put integration response request.

PutMethodInput
Request to add a method to an existing Resource resource.

PutMethodResponseInput
Request to add a MethodResponse to an existing Method resource.

PutRestApiInput
A PUT request to update an existing API, with external API definitions specified as the request body.

TagResourceInput
Adds or updates a tag on a given resource.

TestInvokeAuthorizerInput
Make a request to simulate the execution of an Authorizer.

TestInvokeMethodInput
Make a request to simulate the execution of a Method.

UntagResourceInput
Removes a tag from a given resource.

UpdateAccountInput
Requests API Gateway to change information about the current Account resource.

UpdateApiKeyInput
A request to change information about an ApiKey resource.

UpdateAuthorizerInput
Request to update an existing Authorizer resource.

UpdateBasePathMappingInput
A request to change information about the BasePathMapping resource.

UpdateClientCertificateInput
A request to change information about an ClientCertificate resource.

UpdateDeploymentInput
Requests API Gateway to change information about a Deployment resource.

UpdateDocumentationPartInput
Updates an existing documentation part of a given API.

UpdateDocumentationVersionInput
Updates an existing documentation version of an API.

UpdateDomainNameInput
A request to change information about the DomainName resource.

UpdateGatewayResponseInput
Updates a GatewayResponse of a specified response type on the given RestApi.

UpdateIntegrationInput
Represents an update integration request.

UpdateIntegrationResponseInput
Represents an update integration response request.

UpdateMethodInput
Request to update an existing Method resource.

UpdateMethodResponseInput
A request to update an existing MethodResponse resource.

UpdateModelInput
Request to update an existing model in an existing RestApi resource.

UpdateRequestValidatorInput
Updates a RequestValidator of a given RestApi.

UpdateResourceInput
Request to change information about a Resource resource.

UpdateRestApiInput
Request to update an existing RestApi resource in your collection.

UpdateStageInput
Requests API Gateway to change information about a Stage resource.

UpdateUsageInput
The PATCH request to grant a temporary extension to the remaining quota of a usage plan associated with a specified API key.

UpdateUsagePlanInput
The PATCH request to update a usage plan of a given plan Id.

UpdateVpcLinkInput
Updates an existing VpcLink of a specified identifier.