Module aws_sdk_apigateway::operation

Structs

CreateApiKey
Create an ApiKey resource.

CreateAuthorizer
Adds a new Authorizer resource to an existing RestApi resource.

CreateBasePathMapping
Creates a new BasePathMapping resource.

CreateDeployment
Creates a Deployment resource, which makes a specified RestApi callable over the internet.

CreateDocumentationPart

CreateDocumentationVersion

CreateDomainName
Creates a new domain name.

CreateModel
Adds a new Model resource to an existing RestApi resource.

CreateRequestValidator
Creates a ReqeustValidator of a given RestApi.

CreateResource
Creates a Resource resource.

CreateRestApi
Creates a new RestApi resource.

CreateStage
Creates a new Stage resource that references a pre-existing Deployment for the API.

CreateUsagePlan
Creates a usage plan with the throttle and quota limits, as well as the associated API stages, specified in the payload.

CreateUsagePlanKey
Creates a usage plan key for adding an existing API key to a usage plan.

CreateVpcLink
Creates a VPC link, under the caller's account in a selected region, in an asynchronous operation that typically takes 2-4 minutes to complete and become operational. The caller must have permissions to create and update VPC Endpoint services.

DeleteApiKey
Deletes the ApiKey resource.

DeleteAuthorizer
Deletes an existing Authorizer resource.

DeleteBasePathMapping
Deletes the BasePathMapping resource.

DeleteClientCertificate
Deletes the ClientCertificate resource.

DeleteDeployment
Deletes a Deployment resource. Deleting a deployment will only succeed if there are no Stage resources associated with it.

DeleteDocumentationPart

DeleteDocumentationVersion

DeleteDomainName
Deletes the DomainName resource.

DeleteGatewayResponse
Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

DeleteIntegration
Represents a delete integration.

DeleteIntegrationResponse
Represents a delete integration response.

DeleteMethod
Deletes an existing Method resource.

DeleteMethodResponse
Deletes an existing MethodResponse resource.

DeleteModel
Deletes a model.

DeleteRequestValidator
Deletes a RequestValidator of a given RestApi.

DeleteResource
Deletes a Resource resource.

DeleteRestApi
Deletes the specified API.

DeleteStage
Deletes a Stage resource.

DeleteUsagePlan
Deletes a usage plan of a given plan Id.

DeleteUsagePlanKey
Deletes a usage plan key and remove the underlying API key from the associated usage plan.

DeleteVpcLink
Deletes an existing VpcLink of a specified identifier.

FlushStageAuthorizersCache
Flushes all authorizer cache entries on a stage.

FlushStageCache
Flushes a stage's cache.

GenerateClientCertificate
Generates a ClientCertificate resource.

GetAccount
Gets information about the current Account resource.

GetApiKey
Gets information about the current ApiKey resource.

GetApiKeys
Gets information about the current ApiKeys resource.

GetAuthorizer
Describe an existing Authorizer resource.

GetAuthorizers
Describe an existing Authorizers resource.

GetBasePathMapping
Describe a BasePathMapping resource.

GetBasePathMappings
Represents a collection of BasePathMapping resources.

GetClientCertificate
Gets information about the current ClientCertificate resource.

GetClientCertificates
Gets a collection of ClientCertificate resources.

GetDeployment
Gets information about a Deployment resource.

GetDeployments
Gets information about a Deployments collection.

GetDocumentationPart

GetDocumentationParts

GetDocumentationVersion

GetDocumentationVersions

GetDomainName
Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

GetDomainNames
Represents a collection of DomainName resources.

GetExport
Exports a deployed version of a RestApi in a specified format.

GetGatewayResponse
Gets a GatewayResponse of a specified response type on the given RestApi.

GetGatewayResponses
Gets the GatewayResponses collection on the given RestApi. If an API developer has not added any definitions for gateway responses, the result will be the API Gateway-generated default GatewayResponses collection for the supported response types.

GetIntegration
Get the integration settings.

GetIntegrationResponse
Represents a get integration response.

GetMethod
Describe an existing Method resource.

GetMethodResponse
Describes a MethodResponse resource.

GetModel
Describes an existing model defined for a RestApi resource.

GetModelTemplate
Generates a sample mapping template that can be used to transform a payload into the structure of a model.

GetModels
Describes existing Models defined for a RestApi resource.

GetRequestValidator
Gets a RequestValidator of a given RestApi.

GetRequestValidators
Gets the RequestValidators collection of a given RestApi.

GetResource
Lists information about a resource.

GetResources
Lists information about a collection of Resource resources.

GetRestApi
Lists the RestApi resource in the collection.

GetRestApis
Lists the RestApis resources for your collection.

GetSdk
Generates a client SDK for a RestApi and Stage.

GetSdkType

GetSdkTypes

GetStage
Gets information about a Stage resource.

GetStages
Gets information about one or more Stage resources.

GetTags
Gets the Tags collection for a given resource.

GetUsage
Gets the usage data of a usage plan in a specified time interval.

GetUsagePlan
Gets a usage plan of a given plan identifier.

GetUsagePlanKey
Gets a usage plan key of a given key identifier.

GetUsagePlanKeys
Gets all the usage plan keys representing the API keys added to a specified usage plan.

GetUsagePlans
Gets all the usage plans of the caller's account.

GetVpcLink
Gets a specified VPC link under the caller's account in a region.

GetVpcLinks
Gets the VpcLinks collection under the caller's account in a selected region.

ImportApiKeys
Import API keys from an external source, such as a CSV-formatted file.

ImportDocumentationParts

ImportRestApi
A feature of the API Gateway control service for creating a new API from an external API definition file.

PutGatewayResponse
Creates a customization of a GatewayResponse of a specified response type and status code on the given RestApi.

PutIntegration
Sets up a method's integration.

PutIntegrationResponse
Represents a put integration.

PutMethod
Add a method to an existing Resource resource.

PutMethodResponse
Adds a MethodResponse to an existing Method resource.

PutRestApi
A feature of the API Gateway control service for updating an existing API with an input of external API definitions.

TagResource
Adds or updates a tag on a given resource.

TestInvokeAuthorizer
Simulate the execution of an Authorizer in your RestApi with headers, parameters, and an incoming request body.

TestInvokeMethod
Simulate the execution of a Method in your RestApi with headers, parameters, and an incoming request body.

UntagResource
Removes a tag from a given resource.

UpdateAccount
Changes information about the current Account resource.

UpdateApiKey
Changes information about an ApiKey resource.

UpdateAuthorizer
Updates an existing Authorizer resource.

UpdateBasePathMapping
Changes information about the BasePathMapping resource.

UpdateClientCertificate
Changes information about an ClientCertificate resource.

UpdateDeployment
Changes information about a Deployment resource.

UpdateDocumentationPart

UpdateDocumentationVersion

UpdateDomainName
Changes information about the DomainName resource.

UpdateGatewayResponse
Updates a GatewayResponse of a specified response type on the given RestApi.

UpdateIntegration
Represents an update integration.

UpdateIntegrationResponse
Represents an update integration response.

UpdateMethod
Updates an existing Method resource.

UpdateMethodResponse
Updates an existing MethodResponse resource.

UpdateModel
Changes information about a model.

UpdateRequestValidator
Updates a RequestValidator of a given RestApi.

UpdateResource
Changes information about a Resource resource.

UpdateRestApi
Changes information about the specified API.

UpdateStage
Changes information about a Stage resource.

UpdateUsage
Grants a temporary extension to the remaining quota of a usage plan associated with a specified API key.

UpdateUsagePlan
Updates a usage plan of a given plan Id.

UpdateVpcLink
Updates an existing VpcLink of a specified identifier.